refactor: separating responsabilities
